+++ inst/NEWS.md
@@ -13,6 +13,8 @@ __MEDseq: Mixtures of Exponential-Distance Models with Covariates__
 ### New Features & Improvements
 * `MEDseq_control` gains the arg. `random=TRUE`, governing tie-breaking of estimated central sequence  
   positions: old behaviour (always choosing the first candidate state) recoverable via `random=FALSE`.
+* `plot.MEDseq` arg. `quant.scale=FALSE` replaces old arg. `log.scale`: quantiles now used  
+  to determine non-linear colour breakpoints when invoked with `type="precision"`.
 * Sped-up `init.z="kmedoids"` initialisation via `pam` for _unweighted_ sequences, by using the  
   _highest available_ value for the `pamonce` option,  based on the `cluster` package's version number.
 * `init.z` gains the options `"kmodes"` & `"kmodes2"`, though only for _unweighted_ sequences:  
@@ -20,20 +22,27 @@ __MEDseq: Mixtures of Exponential-Distance Models with Covariates__
 * For weighted sequences, `init.z="kmedoids"` is now itself initialised by Ward's hierarchical clustering.
 * Significant speed-ups to computation of central sequences for all `opti` settings (esp. `"mode"`).
 * Added `SPS` arg. (default=`FALSE`) to `print.MEDtheta` & `summary.MEDseq`.
+* `seriate="observations"` (the default) now also works for `type="I"` plots.
+* `seriate="clusters"` now also works for `type="dbsvals"` & `type="aswvals"` plots.
 * `dbs` gains the optional/experimental arg. `clusters` - no change to default.
   
 ### Bug Fixes & Miscellaneous Edits
 * `MEDseq_fit` now always internally normalises the `weights` to sum to the sample size.
 * Minor fixes to properly account for weighted sequences &/or duplicates when `noise.gate=FALSE`.
+* Minor fix to gathering of results to account for `noise.gate=FALSE` when `G=2`.
 * Fixed `seriate` options `"clusters"`/`"both"` in `plot.MEDseq` for models with no noise component.
 * `MEDseq_stderr` now respects the `algo`, `opti`, & `noise.gate` settings of the original model.
 * `MEDseq_compare` now returns & prints `opti` info where relevant.
 * Fixes to `print` & `summary` methods for `MEDgating` objects if `equalPro=TRUE`.
+* `MEDseq_fit` now coerces `"character"` covariates to `"factor"`.
 * Minor fixes to `print` method for `MEDlambda` objects also.
 * Additional minor edits to `plot.MEDseq(..., type="gating")`.
-* `print.MEDseqCompare` gains the args. `maxi` & `rerank=FALSE`.
+* `print.MEDseqCompare` gains the args. `maxi`, & `rerank=FALSE`.
+* Minor speed-ups for `G=1` models.
+* Added `viridisLite (>= 0.2.0)` to `Suggests:` (for `plot.MEDseq(..., type="precision")`).
 * Ensured `matrixStats (>= 0.53.1)` and `TraMineR (>= 1.6)` in `Imports:`.
 * Package startup message now checks if newer version of package is available from CRAN.
+* Significant vignette edits.
 * Minor documentation & examples edits.
 
 ## MEDseq v1.1.1 - (_4<sup>th</sup> release [patch update]: 2020-05-12_)